// src/models/user.ts
export interface User {
    assign_permission?: String[]
    branch_id?: String[]
    created_at: Date
    email?: String
    email_verified_at?: Date
    id: Number
    login_id?: String
    name?: String
    phone?: String
    role_id?: String
    status: Boolean
    updated_at?: Date
    user_password?: String
}